[directionsbuilder] fixing type

+++ src/odin/directionsbuilder.cc
@@ -67,7 +67,7 @@ void DirectionsBuilder::Build(Api& api, const MarkupFormatter& markup_formatter)
 // Update the heading of ~0 length edges.
 void DirectionsBuilder::UpdateHeading(EnhancedTripLeg* etp) {
 
-  for (size_t x = 0; x < etp->node_size(); ++x) {
+  for (int x = 0; x < etp->node_size(); ++x) {
     auto prev_edge = etp->GetPrevEdge(x);
     auto curr_edge = etp->GetCurrEdge(x);
     auto next_edge = etp->GetNextEdge(x);
